refactor: 提取出共有的useMsgList逻辑
