+++ src/web/components/messageTypes/bbcode/ImageTag.tsx
@@ -3,7 +3,7 @@ import { TagProps } from '@shared/components/bbcode/type';
 import Image from '@web/components/Image';
 import config from '@shared/project.config';
 import ImageViewer from '@web/components/ImageViewer';
-import { useMsgListImageUrls } from '@shared/context/MsgListContext';
+import { useMsgListContextImageUrls } from '@shared/context/MsgListContext';
 
 const ImageTag: React.FC<TagProps> = React.memo((props) => {
   const { node } = props;
@@ -11,7 +11,7 @@ const ImageTag: React.FC<TagProps> = React.memo((props) => {
   let src = node.content.join('');
   src = config.file.getAbsolutePath(src);
 
-  const imageUrls = useMsgListImageUrls();
+  const imageUrls = useMsgListContextImageUrls();
 
   return (
     <ImageViewer originImageUrl={src} allImageUrls={imageUrls}>
